Prevent grpc file clobbering
Due to file name conflicts between prost and tonic generation some
generated files will get 'clobbered'. Meaning prost will generate one
version of the file that does not contain the grpc endpoints and Tonic
will generate a different one that does.

If we wanted only one of the two (for example with cosmos-sdk) we can
simply run Tonic second and be happy. But in the tendermint-proto case
there are several files without any grpc endpoints, this causes tonic to
generate an empty file where prost generates a file of structs.

This leaves us in the awkward position of needing some files generated
by tonic and some files generated by prost and both will clobber some
files we need from the other regardless of the order we run them in.
This patch resolve that issue by renaming specific files we need tonic
endpoints for. We chose to rename the tonic files becuase there are
fewer of them that we need than the prost files.